+++ crates/cli-tools/src/lib.rs
@@ -16,6 +16,7 @@
 //!
 //! This library is also used for the internal maintenance CLI of Wasefire called xtask.
 
+#![feature(async_fn_track_caller)]
 #![feature(path_add_extension)]
 #![feature(try_find)]